Refactor constants according to requirements (#222)
