+++ src/test/java/xyz/npgw/test/common/Constants.java
@@ -1,0 +1,16 @@
+package xyz.npgw.test.common;
+
+public class Constants {
+
+    public static final String LOGIN_PAGE_URL = ProjectProperties.getBaseUrl() + "/";
+    public static final String DASHBOARD_PAGE_URL = ProjectProperties.getBaseUrl() + "/dashboard";
+    public static final String TRANSACTIONS_PAGE_URL = ProjectProperties.getBaseUrl() + "/transactions";
+    public static final String REPORTS_PAGE_URL = ProjectProperties.getBaseUrl() + "/reports";
+    public static final String SA_PAGE_URL = ProjectProperties.getBaseUrl() + "/system";
+
+    public static final String LOGIN_URL_TITLE = "NPGW";
+    public static final String DASHBOARD_URL_TITLE = "NPGW";
+    public static final String TRANSACTIONS_URL_TITLE = "NPGW";
+    public static final String REPORTS_URL_TITLE = "NPGW";
+    public static final String SA_URL_TITLE = "NPGW";
+}
