+++ src/test/java/xyz/npgw/test/page/systemadministration/CompaniesAndBusinessUnitsPage.java
@@ -9,7 +9,7 @@ import xyz.npgw.test.page.base.SystemAdministrationWithTableBasePage;
 
 public class CompaniesAndBusinessUnitsPage extends SystemAdministrationWithTableBasePage {
 
-    private final Locator addCompanyButton = locator("svg[data-icon='circle-plus']").first();
+    private final Locator addCompanyButton = locator("button[data-testid='AddCompanyButton']");
     @Getter
     private final Locator addCompanyDialog = dialog();
