+++ src/test/java/xyz/npgw/test/run/AddCompanyDialogTest.java
@@ -7,10 +7,10 @@ import io.qameta.allure.Feature;
 import io.qameta.allure.TmsLink;
 import org.testng.annotations.Test;
 import xyz.npgw.test.common.base.BaseTest;
+import xyz.npgw.test.common.provider.TestDataProvider;
 import xyz.npgw.test.page.AddCompanyDialog;
 import xyz.npgw.test.page.DashboardPage;
 import xyz.npgw.test.page.systemadministration.CompaniesAndBusinessUnitsPage;
-import xyz.npgw.test.testdata.TestDataProvider;
 
 import java.util.List;
