+++ src/test/java/xyz/npgw/test/run/LoginPageTest.java
@@ -7,11 +7,12 @@ import io.qameta.allure.Feature;
 import io.qameta.allure.TmsLink;
 import org.testng.annotations.Optional;
 import org.testng.annotations.Test;
+import xyz.npgw.test.common.Constants;
+import xyz.npgw.test.common.ProjectProperties;
 import xyz.npgw.test.common.base.BaseTest;
+import xyz.npgw.test.common.provider.TestDataProvider;
 import xyz.npgw.test.page.AboutBlankPage;
 import xyz.npgw.test.page.LoginPage;
-import xyz.npgw.test.testdata.Constants;
-import xyz.npgw.test.testdata.TestDataProvider;
 
 import static com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat;
 
@@ -41,15 +42,15 @@ public class LoginPageTest extends BaseTest {
     public void testRememberMeCheckedSavesUserEmail(@Optional("GUEST") String userRole) {
         LoginPage loginPage = new AboutBlankPage(getPage())
                 .navigate("/login")
-                .fillEmailField(Constants.USER_EMAIL)
-                .fillPasswordField(Constants.USER_PASSWORD)
+                .fillEmailField(ProjectProperties.getUserEmail())
+                .fillPasswordField(ProjectProperties.getUserPassword())
                 .checkRememberMeCheckbox()
                 .clickLoginButton()
                 .getHeader()
                 .clickLogOutButton();
 
         Allure.step("Verify: The user's email is in the email field");
-        assertThat(loginPage.getEmailField()).hasValue(Constants.USER_EMAIL);
+        assertThat(loginPage.getEmailField()).hasValue(ProjectProperties.getUserEmail());
     }
 
     @Test
@@ -60,8 +61,8 @@ public class LoginPageTest extends BaseTest {
     public void testRememberMeUncheckedDontSaveUserEmail(@Optional("GUEST") String userRole) {
         LoginPage loginPage = new AboutBlankPage(getPage())
                 .navigate("/")
-                .fillEmailField(Constants.USER_EMAIL)
-                .fillPasswordField(Constants.USER_PASSWORD)
+                .fillEmailField(ProjectProperties.getUserEmail())
+                .fillPasswordField(ProjectProperties.getUserPassword())
                 .uncheckRememberMeCheckbox()
                 .clickLoginButton()
                 .getHeader()
