+++ src/test/java/xyz/npgw/test/run/ReportsPageTest.java
@@ -6,10 +6,10 @@ import io.qameta.allure.Epic;
 import io.qameta.allure.Feature;
 import io.qameta.allure.TmsLink;
 import org.testng.annotations.Test;
+import xyz.npgw.test.common.Constants;
 import xyz.npgw.test.common.base.BaseTest;
 import xyz.npgw.test.page.DashboardPage;
 import xyz.npgw.test.page.ReportsPage;
-import xyz.npgw.test.testdata.Constants;
 
 import static com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat;
